Massive update - Switched parsing to use an AST - no syntax changes
Just about everything related to parsing and executing the script has
been re-written. Only some of the JobRunner code has been left intact.

Instead of parsing the file line by line, and executing it as we go, 
now we parse the script into an Abstract Syntax Tree (of sorts). After
the AST has been computed, then it is executed. This will setup the
build-targets. After this, *then* we try to eval/run build-targets. All
of the key statements (if, for, and build-target definitions) are
implemented as nodes in the AST. This way we can define build-targets at
runtime, including parsable outputs... for example:

for i in 1..5:
	if i != 3
	file.${i} : 
		touch $>
	endif

this will now create separate build-target definitions for: file.1,
file.2, file.4, and file.5. This is different from something like:

file.@{1..5} : 

which is one build-target that will create all of the files at once.