added support for opportunistic jobs
these are jobs that will always run, even if they aren't directly
called.

However, they will not submit dependencies. Their inputs must either a)
exists on disk or b) be part of an already submitted job.

The main use case for this is to remove temporary files, but only after
the main output has been generated. The reason to use this over the
simplier '^tmpfile: input1 input2' specification is to protect the temp$
files in the case of a job failure.

If a job fails and gets resubmitted, the temp file list (in cgpipe) is different on
the resubmission.