Complex dependency job submission with temp files
The basic idea is that if a final output exists, but the earlier parent
files do not, then we should re-build the pipeline to get a fresh final
output. However, if the earlier files are marked as "temporary"
(^fname), then we can skip rebuilding the temp intermediate files, so
long as the final output file exists.