Added file lock for job-log, added update-job-end/start for cgpipe
Job log locks are needed to avoid a race condition when reading/writing to a job-log. Lock is based on creating
and removing a lock directory. This is the only atomic way for locking on *nix systems.

There are two related commands added for cgpipe: update-job-start and update-job-end. These will update the job-log
as necessary to add this information.