BUGFIX lists weren't being dereferenced in methods
Previously, if a list was used with a method (ex: mylist[1].upper()),
then the line would fail (infinte loop) if the line wasn't a simple
evaluation. If the line had an operators (tmp = foo.split()[1].upper()),
then the parser would fail.

Added test to reproduce.

also migrated a few read-only vars to the RootContext (ones that don't
depend on the exact pipeline script being executed.