Skippability of targets depends on downstream jobs
If a temp output is missing -- it will be re-created *if* it is needed
for a downstream file.

imagine this setup:

^temp1: input1
   script

output1: temp1
    script.


if output1 does not exist and temp1 does not exist, then temp1 will be
built. Otherwise, if output1 exists, then we don't need to worry about
temp1...